requirements.txt
streamlit
langchain
langchain-core
requests
streamlit-lottie
python-dotenv
openai
-U langchain-openai